+++ src/main/java/appeng/client/gui/implementations/GuiWirelessInterfaceConfigurationTerminal.java
@@ -18,6 +18,7 @@ import java.util.List;
 import net.minecraft.client.gui.GuiButton;
 import net.minecraft.entity.player.InventoryPlayer;
 
+import appeng.client.gui.widgets.GuiSlotIndicator;
 import appeng.client.gui.widgets.GuiTerminalModeSwitch;
 import appeng.client.gui.widgets.GuiWirelessUpgradePlate;
 import appeng.container.implementations.ContainerWirelessInterfaceConfigurationTerminal;
@@ -81,6 +82,12 @@ public class GuiWirelessInterfaceConfigurationTerminal extends GuiInterfaceConfi
         GuiWirelessUpgradePlate.draw(this, offsetX + PLATE_X, offsetY + this.plateY(),
                 IWirelessTerminalContainer.UPGRADE_SLOTS);
         super.drawBG(offsetX, offsetY, mouseX, mouseY);
+    }
+
+    @Override
+    public void drawFG(final int offsetX, final int offsetY, final int mouseX, final int mouseY) {
+        super.drawFG(offsetX, offsetY, mouseX, mouseY);
+        GuiSlotIndicator.draw(this);
     }
 
     @Override